fixed configuration in knexfile.js for proper local postgres set up. Also moved knexfile.js from config to root folder because the path wasn't resolving correctly when calling knex commands from terminal.
