+++ data/db-config.js
@@ -1,5 +1,5 @@
 require('dotenv').config();
 const knex = require('knex');
 const dbEnvironment = process.env.NODE_ENV || 'development';
-const knexConfig = require('../config/knexfile')[dbEnvironment];
+const knexConfig = require('../knexfile')[dbEnvironment];
 module.exports = knex(knexConfig);
