+++ package.json
@@ -59,7 +59,7 @@
     "faker": "^4.1.0",
     "helmet": "^3.23.1",
     "http-errors": "~1.6.3",
-    "knex": "^0.21.1",
+    "knex": "^0.21.16",
     "knex-cleaner": "^1.3.1",
     "morgan": "~1.9.1",
     "pg": "^8.2.1",
